convert roles stores to typescript
